azstoragepy/requirements.txt
azure-core==1.6.0
azure-storage-blob==12.3.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.10
isodate==0.6.0
msrest==0.6.17
oauthlib==3.1.0
pycparser==2.20
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.9
